Fix for systems that contain ChBodyAuxRef:
In order to properly deal with ChBodyAuxRef it is *important* to call the Chrono virtual function update().  For a ChBody, this function performs all operations that were explicitly performed in the chrono_parallel code.  However, for a ChBodyAuxRef this function is overriden: it first calls the base class (ChBody) update then performs the required update for the noncentroidal reference frame.